Move `getTabCompletionCoordinateExact` to `CommandCarpetBase`
