+++ carpetmodSrc/carpet/commands/CommandCarpetBase.java
@@ -7,12 +7,79 @@ import net.minecraft.command.ICommandSender;
 import net.minecraft.command.NumberInvalidException;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.MathHelper;
+import net.minecraft.util.math.Vec3d;
 import net.minecraft.util.text.ITextComponent;
 
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
+import java.util.ArrayList;
 import java.util.List;
 
 public abstract class CommandCarpetBase extends CommandBase
 {
+    @Nonnull
+    @ParametersAreNonnullByDefault
+    public static List<String> getTabCompletionCoordinateExact(ICommandSender sender, String[] inputArgs, int index, @Nullable BlockPos target, int decimals) {
+        Vec3d posBaseE = sender.getPositionVector();
+        List<String> list = new ArrayList<>();
+        int i = inputArgs.length - 1;
+        double posCurrE;
+        int posCurrB;
+        int posTarget;
+        switch (i - index) {
+            case 0:
+                posCurrE = posBaseE.x;
+                posCurrB = MathHelper.floor(posCurrE);
+                if (target != null) {
+                    posTarget = target.getX();
+                } else {
+                    posTarget = posCurrB;
+                }
+                break;
+            case 1:
+                posCurrE = posBaseE.y;
+                posCurrB = MathHelper.floor(posCurrE);
+                if (target != null) {
+                    posTarget = target.getY();
+                } else {
+                    posTarget = posCurrB;
+                }
+                break;
+            case 2:
+                posCurrE = posBaseE.z;
+                posCurrB = MathHelper.floor(posCurrE);
+                if (target != null) {
+                    posTarget = target.getZ();
+                } else {
+                    posTarget = posCurrB;
+                }
+                break;
+            default:
+                return list;
+        }
+        list.add("~");
+        if (decimals > 0 && decimals < 8) {
+            String formatter = "%." + decimals + 'f';
+            list.add(String.format(formatter, posCurrE));
+        }
+        list.add(Double.toString(posCurrE));
+        list.add(Integer.toString(posCurrB));
+        if (posTarget != posCurrB) {
+            list.add(Integer.toString(posTarget));
+        }
+        return list;
+    }
+
+    @SuppressWarnings("unused")
+    @Nonnull
+    @ParametersAreNonnullByDefault
+    public static List<String> getTabCompletionCoordinateExact(ICommandSender sender, String[] inputArgs, int index, @Nullable BlockPos target) {
+        return CommandCarpetBase.getTabCompletionCoordinateExact(sender, inputArgs, index, target, -1);
+    }
+
     @Override
     public boolean checkPermission(MinecraftServer server, ICommandSender sender) {
         return true;
