+++ carpetmodSrc/carpet/commands/CommandPerimeterInfo.java
@@ -9,7 +9,6 @@ import net.minecraft.entity.EntityLiving;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.DimensionType;
 import net.minecraft.world.World;
@@ -17,72 +16,11 @@ import net.minecraft.world.World;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
 public class CommandPerimeterInfo extends CommandCarpetBase {
     private static final String USAGE = "/perimeterinfo <x> <y> <z> [<dimension> [<target_entity>]]";
-
-    @Nonnull
-    @ParametersAreNonnullByDefault
-    public static List<String> getTabCompletionCoordinateExact(ICommandSender sender, String[] inputArgs, int index, @Nullable BlockPos target, int decimals) {
-        Vec3d posBaseE = sender.getPositionVector();
-        List<String> list = new ArrayList<>();
-        int i = inputArgs.length - 1;
-        double posCurrE;
-        int posCurrB;
-        int posTarget;
-        switch (i - index) {
-            case 0:
-                posCurrE = posBaseE.x;
-                posCurrB = MathHelper.floor(posCurrE);
-                if (target != null) {
-                    posTarget = target.getX();
-                } else {
-                    posTarget = posCurrB;
-                }
-                break;
-            case 1:
-                posCurrE = posBaseE.y;
-                posCurrB = MathHelper.floor(posCurrE);
-                if (target != null) {
-                    posTarget = target.getY();
-                } else {
-                    posTarget = posCurrB;
-                }
-                break;
-            case 2:
-                posCurrE = posBaseE.z;
-                posCurrB = MathHelper.floor(posCurrE);
-                if (target != null) {
-                    posTarget = target.getZ();
-                } else {
-                    posTarget = posCurrB;
-                }
-                break;
-            default:
-                return list;
-        }
-        list.add("~");
-        if (decimals > 0 && decimals < 8) {
-            String formatter = "%." + decimals + 'f';
-            list.add(String.format(formatter, posCurrE));
-        }
-        list.add(Double.toString(posCurrE));
-        list.add(Integer.toString(posCurrB));
-        if (posTarget != posCurrB) {
-            list.add(Integer.toString(posTarget));
-        }
-        return list;
-    }
-
-    @SuppressWarnings("unused")
-    @Nonnull
-    @ParametersAreNonnullByDefault
-    public static List<String> getTabCompletionCoordinateExact(ICommandSender sender, String[] inputArgs, int index, @Nullable BlockPos target) {
-        return getTabCompletionCoordinateExact(sender, inputArgs, index, target, -1);
-    }
 
     @Override
     @Nonnull
